requirements-airflow.txt
kafka-python==2.3.0
requests==2.32.3
python-dotenv==1.0.0
google-cloud-bigquery==3.40.1